Linux build working.
Had to change vector constructors because g++ compiler does not
like constructors in anonymous unions. So to make a vector you now
write
    vec2 x = V2(1, 2);
instead of
    vec2 x = vec2(1, 2);

Changed the offset parameter in attribv from a u32 to void*.
It is unlikely that you will have a need to indicate an offset
greater than 2^32, but hey, g++ was complaining.